Let an AutoUX editor accept more than one type for a value
The AutoUX editors already took `USSType[]`, but everything downstream of the
selector quietly used `type[0]`: switching to "Custom Expression" recorded only
the first type as the expected type, and picking "Constant" built a default of
the first type whether or not a constant could have that type. With every call
site passing a single-element array that was invisible. The condition editor
needs a value that is either a number or a number vector, so it stops being
invisible there.

`possibilities` had the matching restriction: it offered constants *or*
variables and functions, never both, because with one type those cases are
mutually exclusive. It now offers constants when some type can be a constant
and variables when some type cannot, which is the same answer for every
existing caller.

The default-expression helpers move to `parseExpr.ts` alongside the parsing
they mirror, so that code building a default value doesn't have to import a
React component to do it.
